Add common base class for publication and subscription listeners
