+++ tools/rtpsrelay/PublicationListener.h
@@ -2,15 +2,15 @@
 #define RTPSRELAY_PUBLICATION_LISTENER_H_
 
 #include "RelayStatisticsReporter.h"
+#include "GuidAddrSet.h"
 #include "GuidPartitionTable.h"
-#include "ReaderListenerBase.h"
-#include "RelayHandler.h"
+#include "EndpointListener.h"
 
 #include <dds/DCPS/DomainParticipantImpl.h>
 
 namespace RtpsRelay {
 
-class PublicationListener : public ReaderListenerBase {
+class PublicationListener : public EndpointListener {
 public:
   PublicationListener(const Config& config,
                       const GuidAddrSet_rch& guid_addr_set,
@@ -22,9 +22,7 @@ private:
   void on_data_available(DDS::DataReader_ptr reader) override;
 
   const Config& config_;
-  GuidAddrSet_rch guid_addr_set_;
   OpenDDS::DCPS::DomainParticipantImpl* participant_;
-  GuidPartitionTable& guid_partition_table_;
   RelayStatisticsReporter& stats_reporter_;
   OpenDDS::DCPS::Atomic<size_t> count_;
 };
